Standardize on doozer working dir
This commit removes `--working-dir` from `doozer` invocations, and sets
the working dir through the environment variable `DOOZER_WORKING_DIR`
instead.

When `buildlib` is loaded.

As a side effect, the build logs will be a tiny bit more readable, and
`build-scripts/find-and-compress-brew-logs.sh` will not take 4 minutes
anymore to find the directory for brew logs on a job with a busy build,
as it already knows where the directory is.